ETDA-1665 Moved author submit final submission logic into FinalSubmissionSubmitService.
